// src/domain/enums/StatusRastreamento.ts
export enum StatusRastreamento {
  AGUARDANDO_TRIAGEM = "AGUARDANDO_TRIAGEM",
  EM_TRIAGEM = "EM_TRIAGEM",
  AGUARDANDO_DESMONTE = "AGUARDANDO_DESMONTE",
  EM_DESMONTE = "EM_DESMONTE",
  PECAS_REAPROVEITADAS = "PECAS_REAPROVEITADAS",
  MATERIAL_RECICLAVEL = "MATERIAL_RECICLAVEL",
  DESCARTE_SEGURO = "DESCARTE_SEGURO",
  BAIXA_DEFINITIVA = "BAIXA_DEFINITIVA"
}